feat: Rename module to Drops
