+++ Readme.md
@@ -1,3 +1,4 @@
-# How to test - temporary solution  
+# Drops
+  
 mark a bunch of prefabs using "Lootable"  
 use `giveRandomItem` ingame lot of times
